requirements.txt
tqdm~=4.66.4
absl-py
torch
transformers~=4.41.2
langchain~=0.2.6
langchain-community
langchain-openai
langchain-huggingface
gradio
unstructured
pdf2image
pdfminer
pdfminer.six
opencv-python
unstructured_inference
layoutparser
onnxruntime
pikepdf
sentence-transformers
chromadb
pillow_heif
unstructured.pytesseract
pandas~=2.2.2
huggingface-hub~=0.23.4
matplotlib~=3.9.0